[accel-web] refactor
