+++ packages/accel-web/src/form/collectionRadioButtons.astro
@@ -1,6 +1,6 @@
 ---
-import input from "../nativeComponents/input.astro";
-import label from "../nativeComponents/label.astro";
+import input from "../form/input.astro";
+import label from "../form/label.astro";
 import { ex } from "../paginate";
 
 interface Props {
