+++ packages/accel-web/src/index.ts
@@ -1,3 +1,3 @@
-export { formWith } from "./formWith.js";
+export { formFor } from "./form/index.js";
 export { paginate } from "./paginate/index.js";
 export { Parameters } from "./parameters.js";
